Rename org.example -> org.canvas.server
