+++ server/pom.xml
@@ -4,9 +4,9 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>org.example</groupId>
-    <artifactId>actualMavenTest</artifactId>
-    <version>1.0-SNAPSHOT</version>
+    <groupId>org.canvas.server</groupId>
+    <artifactId>canvas-server</artifactId>
+    <version>0.1.0</version>
 
     <properties>
         <java.version>19</java.version>
