+++ server/src/main/java/org/canvas/server/Data.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.canvas.server;
 
 public class Data implements DataTransferObject {
     private long timestamp;
