+++ server/src/main/java/org/canvas/server/DataAccessObject.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.canvas.server;
 
 import java.sql.*;
