+++ server/src/main/java/org/canvas/server/DataTransferObject.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.canvas.server;
 
 public interface DataTransferObject {
