+++ server/src/main/java/org/canvas/server/Key.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.canvas.server;
 
 import java.util.ArrayList;
 import java.util.List;
