+++ server/src/main/java/org/canvas/server/Main.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.canvas.server;
 
 public class Main {
     public static void main(String[] args) {
