+++ server/src/main/java/org/canvas/server/dataDAO.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.canvas.server;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
